+++ CHANGELOG.md
@@ -6,6 +6,7 @@
     * Moves that needs to charge 1 turn to use will have longer duration.
     * It is designed to be finished in 1 second so it won't influence most players.
 * Custom type effectiveness: Use a json file to give an entity specific weaknesses and resistances. For modpack creators & datapack creators to create more interactions with the mobs from the other mods. (Documentation is coming soon.)
+* Pokemon with Klutz ability will no longer use Held Item.
 ### Changes:
 * Changed some of the particle when using moves with specific types.
 * Slightly adjusted the projectile created for the special move that doesn't have special effect currently to make it more accurate.